requirements.txt
discord.py
pyairtable